backend/requirements.txt
flask
flask-cors
flask-pymongo
flask-jwt-extended
python-dotenv
werkzeug
openai
